CR folder
Deleted those files which might be required for hosting on cubehost. This folder contains files created by me and might not be able to host on cubehost. I could not upload the real folder because it had more than 9000 files and on GitHub one can upload only upto 100 files in a folder at once. So, for hosting purposes i am also uploading the zipped file named cr.